# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MechAssist API",
        "description": "API REST : clients, mécaniciens, demandes d’intervention, chat, géolocalisation, notes. Auth : Laravel Sanctum (Bearer token).",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/auth/google": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "operationId": "googleLogin",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_token"
                                ],
                                "properties": {
                                    "id_token": {
                                        "type": "string"
                                    },
                                    "role": {
                                        "type": "string",
                                        "nullable": true,
                                        "enum": [
                                            "client",
                                            "mecanicien"
                                        ]
                                    },
                                    "fcm_token": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK + token"
                    },
                    "401": {
                        "description": "Jeton Google invalide"
                    },
                    "503": {
                        "description": "Google non configuré"
                    }
                },
                "servers": [
                    {
                        "url": "/",
                        "description": "Serveur d’exécution (ex. http://127.0.0.1:8000)"
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token Sanctum : Authorization: Bearer {token}",
                "scheme": "bearer",
                "bearerFormat": "opaque"
            }
        }
    },
...
}